Сommit

+++ src/Main.java
@@ -13,7 +13,7 @@ public class Main {
 
     private static void calcMoney(int moneyYear) {   //Метод для вычесления отпускных
         int DAYSFREE = 28;  // Колличество дней в отпуске
-        double MONTH = 12;
+        int MONTH = 12;
         double DAYS = 29.5;   //Коллличество дней усстановленых законом
         double calc = (moneyYear / MONTH / DAYS)* DAYSFREE;
         String string = String.format ("%.0f", calc);
